requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cryptography==41.0.3
Deprecated==1.2.14
Flask==2.3.2
Flask-JWT-Extended==4.5.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
paypalrestsdk==1.13.1
pycparser==2.21
pydantic==1.10.12
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
typing-extensions==4.7.1
urllib3==2.0.4
vonage==3.8.0
vonage-jwt==1.0.0
werkzeug==2.3.7
wrapt==1.15.0
zipp==3.16.2
